209: final touches

+++ api-rs/src/main/java/net/runelite/rs/api/RSClient.java
@@ -1484,10 +1484,10 @@ public interface RSClient extends RSGameEngine, Client
 	int getRndHue();
 
 	@Import("Tiles_underlays")
-	byte[][][] getTileUnderlays();
+	short[][][] getTileUnderlays();
 
 	@Import("Tiles_overlays")
-	byte[][][] getTileOverlays();
+	short[][][] getTileOverlays();
 
 	@Import("Tiles_shapes")
 	byte[][][] getTileShapes();
